+++ source/isNotEmpty.js
@@ -1,0 +1,5 @@
+import { isEmpty } from './isEmpty.js';
+
+export function isNotEmpty(input) {
+	return !isEmpty(input);
+}