+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+9.3.0
+
+- Breaking change in relation to TS typings of `R.assoc`, `R.dissoc` and `R.modify` - https://github.com/ramda/types/pull/37
+
+- Add `R.isNotEmpty` as it is new method in `Ramda`
+
+- Fix `R.head`/`R.last` TS definition - It returns `undefined` if array has length of 0. Before 
+
 9.2.1
 
 - Broken `Deno` build - [Issue #731](https://github.com/selfrefactor/rambda/issues/731)